Flip order of intersection test.
Previously, the intersection test would run through the feature's
geometries in order and terminate early, whereas the behaviour of
the 'first' intersection we want is the first *admin area* and the
order of the geometries within the feature shouldn't matter.

Also split the `updater` from the `intersects_iterator` in an
attempt to separate policy from the intersection test.