Allow ignoring layers to control recursion.
When `z > maskLevel`, the bulk exporter will generate the child
tiles of a particular tile only if that tile is non-empty. It does
this by looking at whether it "painted", which means it generated
some features in the output.

However, this doesn't work all that well when it detects the ocean
polygon as "painted" and recurses into boring ocean. This patch
allows the user to specify which layers to ignore when making the
decision about whether to generate child tiles.