Abstracted HTTP server request handling.
Previously, the server was hard-coded to make vector tiles. Now,
a `handler_factory` object is passed in with the responsibility of
constructing `request_handler` objects on each thread. The
`request_handler` objects are then called on each thread to
process requests.

This abstraction allows for mocking the responses so that various
exceptional or unusual states can be explored in the test cases.